Split VLM prompt cache planning
